+++ examples/lambda-example/src/main/java/org/platformlambda/example/MainApp.java
@@ -47,7 +47,6 @@ public class MainApp implements EntryPoint {
         // You can create a microservice as a lambda function inline or write it as a regular Java class
         LambdaFunction echo = (headers, body, instance) -> {
             log.info("echo @"+instance+" received - "+headers+", "+body);
-
             Map<String, Object> result = new HashMap<>();
             result.put("headers", headers);
             result.put("body", body);